@@ -1,24 +1,19 @@
 #!/usr/bin/env python3
 from typing import Optional
 
-from cdedb.script import make_backend, setup, Script, CoreBackend
 import cdedb.validationtypes as vtypes
-from cdedb.backend.common import affirm_validation_typed as affirm
-
-# Configuration
-
-# The admin id will need to be replaces before use.
-executing_admin_id = -1
-rs = setup(persona_id=executing_admin_id, dbuser="cdb_admin",
-           dbpassword="9876543210abcdefghijklmnopqrst")()
+from cdedb.backend.common import affirm_validation as affirm
+from cdedb.script import Script
 
 # Prepare stuff
 
-core: CoreBackend = make_backend("core")
+script = Script(dbuser="cdb_persona")
+rs = script.rs()
+core = script.make_backend("core")
 
 # Execution
 
-with Script(rs, dry_run=False):
+with script:
     persona_id: Optional[int] = -1
     while True:
         persona_id = core.next_persona(
@@ -30,7 +25,7 @@ with Script(rs, dry_run=False):
         # Validate ml data
         persona = core.get_ml_user(rs, persona_id)
         try:
-            affirm(vtypes.Persona, persona, _ignore_warnings=True)
+            affirm(vtypes.Persona, persona)
         except Exception as e:
             print("-" * 80)
             print(f"Error for persona {persona_id}:")
@@ -44,7 +39,7 @@ with Script(rs, dry_run=False):
 
         persona = core.get_event_user(rs, persona_id)
         try:
-            affirm(vtypes.Persona, persona, _ignore_warnings=True)
+            affirm(vtypes.Persona, persona)
         except Exception as e:
             print("-" * 80)
             print(f"Error for persona {persona_id}:")
@@ -58,7 +53,7 @@ with Script(rs, dry_run=False):
 
         persona = core.get_total_persona(rs, persona_id)
         try:
-            affirm(vtypes.Persona, persona, _ignore_warnings=True)
+            affirm(vtypes.Persona, persona)
         except Exception as e:
             print("-" * 80)
             print(f"Error for persona {persona_id}:")
