@@ -7,7 +7,4 @@ files and provides python code encapsulating our :py:mod:`psycopg` usage.
 """
 
 #: all available database roles
-DATABASE_ROLES = ("cdb_anonymous", "cdb_persona", "cdb_member",
-                  "cdb_core_admin", "cdb_cde_admin", "cdb_event_admin",
-                  "cdb_ml_admin", "cdb_assembly_admin", "cdb_files_admin",
-                  "cdb_admin")
+DATABASE_ROLES = ("cdb_anonymous", "cdb_persona", "cdb_member", "cdb_admin")
