@@ -10,7 +10,9 @@ their symbolic names provided by this module should be used.
 import enum
 from typing import Dict
 
-from cdedb.subman.machine import SubscriptionAction, SubscriptionState  # pylint: disable=unused-import # noqa
+from cdedb.subman.machine import (  # pylint: disable=unused-import # noqa: F401
+    SubscriptionAction, SubscriptionState,
+)
 
 
 def n_(x: str) -> str:
@@ -185,10 +187,15 @@ class MailinglistDomain(enum.IntEnum):
 
     testmail = 100
 
-    def __str__(self) -> str:
+    def get_domain(self) -> str:
+        """Return the actual domain for this enum member."""
         if self not in _DOMAIN_STR_MAP:
             raise NotImplementedError(n_("This domain is not supported."))
         return _DOMAIN_STR_MAP[self]
+
+    def display_str(self) -> str:
+        """Return a readable string representation to be displayed in the UI."""
+        return self.get_domain()
 
 
 # Instead of importing this, call str() on a MailinglistDomain.
@@ -385,10 +392,12 @@ class AssemblyLogCodes(enum.IntEnum):
     assembly_presider_removed = 36  #:
     attachment_added = 40  #:
     attachment_removed = 41  #:
-    attachment_changed = 42
-    attachment_version_added = 50
-    attachment_version_removed = 51
-    attachment_version_changed = 52
+    attachment_changed = 42  #:
+    attachment_ballot_link_created = 43  #:
+    attachment_ballot_link_deleted = 44  #:
+    attachment_version_added = 50  #:
+    attachment_version_removed = 51  #:
+    attachment_version_changed = 52  #:
 
 
 @enum.unique
