@@ -1,5 +1,7 @@
+"""subman – Powerful Python package to manage subscriptions"""
+
 from .exceptions import SubscriptionError, SubscriptionInfo
-from .machine import SubscriptionAction, SubscriptionState, SubscriptionPolicy
+from .machine import SubscriptionAction, SubscriptionPolicy, SubscriptionState
 from .subman import SubscriptionManager
 
 __all__ = ['SubscriptionAction', 'SubscriptionError', 'SubscriptionInfo',
