@@ -45,8 +45,11 @@ ASSEMBLY_BACKEND_LOG = _LOG_ROOT / "backend-assembly.log"
 WORKER_LOG = _LOG_ROOT / "frontend-worker.log"
 MAILMAN_LOG = _LOG_ROOT / "frontend-mailman.log"
 
+if pathlib.Path('/CONTAINER').is_file():
+    # ldap host server differs for vms and docker containers
+    LDAP_HOST = "ldap"
+
 if CDEDB_TEST:
-    DB_PORT = 6432
     CDB_DATABASE_NAME = os.environ['CDEDB_TEST_DATABASE']
     SERVER_NAME_TEMPLATE = "test_{}_server"
     STORAGE_DIR = pathlib.Path(os.environ['CDEDB_TEST_TMP_DIR'], 'storage')
