@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
-import copy
 import datetime
 import decimal
 
@@ -8,8 +8,8 @@ import pytz
 
 import cdedb.database.constants as const
 from cdedb.common import (
-    CdEDBObject, PERSONA_CDE_FIELDS, PERSONA_CORE_FIELDS, PERSONA_EVENT_FIELDS,
-    QuotaException, CdEDBLog
+    PERSONA_CDE_FIELDS, PERSONA_CORE_FIELDS, PERSONA_EVENT_FIELDS, CdEDBLog,
+    QuotaException,
 )
 from cdedb.query import Query, QueryOperators, QueryScope
 from tests.common import USER_DICT, BackendTest, as_users, nearly_now
@@ -20,14 +20,14 @@ class TestCdEBackend(BackendTest):
 
     @as_users("berta", "vera")
     def test_basics(self) -> None:
-        data = self.core.get_cde_user(self.key,self.user['id'])
+        data = self.core.get_cde_user(self.key, self.user['id'])
         data['display_name'] = "Zelda"
         setter = {k: v for k, v in data.items() if k in
                   {'id', 'display_name', 'telephone'}}
-        generation = self.core.changelog_get_generation(self.key,self.user['id'])
+        generation = self.core.changelog_get_generation(self.key, self.user['id'])
         num = self.core.change_persona(self.key, setter, generation, change_note='note')
         self.assertEqual(1, num)
-        new_data = self.core.get_cde_user(self.key,self.user['id'])
+        new_data = self.core.get_cde_user(self.key, self.user['id'])
         self.assertEqual(data, new_data)
 
     @as_users("berta")
@@ -50,7 +50,8 @@ class TestCdEBackend(BackendTest):
         # ID and not counted in general.
         self.core.get_cde_users(self.key, (1, 6))
 
-        query = Query(scope=QueryScope.cde_member, spec=QueryScope.cde_member.get_spec(),
+        query = Query(scope=QueryScope.cde_member,
+                      spec=QueryScope.cde_member.get_spec(),
                       fields_of_interest=["id"], constraints=[], order=[])
         with self.assertRaises(QuotaException):
             self.cde.submit_general_query(self.key, query)
@@ -120,7 +121,8 @@ class TestCdEBackend(BackendTest):
                 ("country,country2", QueryOperators.empty, None)],
             order=(("family_name,birth_name", True),),)
         result = self.cde.submit_general_query(self.key, query)
-        self.assertEqual({6, 9, 15, 100}, {e['id'] for e in result})
+        expectation = {6, 9, 15, 100}
+        self.assertEqual({e['id'] for e in result}, expectation)
 
     @as_users("vera")
     def test_user_search(self) -> None:
@@ -133,7 +135,7 @@ class TestCdEBackend(BackendTest):
                 ("birthday", QueryOperators.less, datetime.datetime.now())],
             order=(("family_name", True),),)
         result = self.cde.submit_general_query(self.key, query)
-        self.assertEqual({2, 3, 4, 6, 7, 13, 15, 16, 22, 23, 27, 32, 100},
+        self.assertEqual({2, 3, 4, 6, 7, 13, 15, 16, 22, 23, 27, 32, 37, 100},
                          {e['id'] for e in result})
 
     @as_users("vera")
@@ -142,7 +144,7 @@ class TestCdEBackend(BackendTest):
             scope=QueryScope.cde_user,
             spec=QueryScope.cde_user.get_spec(),
             fields_of_interest=("personas.id", "family_name",
-                                   "birthday"),
+                                "birthday"),
             constraints=[
                 ("given_names", QueryOperators.match, 'Berta'),
                 ("address", QueryOperators.oneof, ("Auf der Düne 42", "Im Garten 77")),
@@ -167,7 +169,7 @@ class TestCdEBackend(BackendTest):
 
     @as_users("vera")
     def test_demotion(self) -> None:
-        self.assertLess(0, self.core.change_membership(self.key, 2, False))
+        self.assertLess(0, self.cde.change_membership(self.key, 2, False)[0])
 
     @as_users("farin")
     def test_lastschrift(self) -> None:
