@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import unittest
 
@@ -17,4 +18,5 @@ class TestBackendCommon(unittest.TestCase):
         self.assertTrue(callable(proxy.login))
         self.assertTrue(callable(proxy.get_realms_multi))
         with self.assertRaises(PrivilegeError):
-            _ = proxy.verify_password
+            # pylint: disable=pointless-statement
+            proxy.verify_password  # exception in __getitem__
