@@ -1,23 +1,23 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import copy
 import datetime
 import decimal
-from typing import cast, Optional
+from typing import Optional, cast
 
 import cdedb.database.constants as const
-from cdedb.common import (
-    CdEDBObject, PERSONA_CDE_FIELDS, PERSONA_EVENT_FIELDS, PERSONA_ML_FIELDS,
-    ArchiveError, PrivilegeError, RequestState, get_hash, merge_dicts, now, nearly_now
-)
-from cdedb.backend.common import affirm_validation_typed as affirm
-from cdedb.validation import PERSONA_CDE_CREATION
 import cdedb.validationtypes as vtypes
-from tests.common import (
-    ANONYMOUS, BackendTest, USER_DICT, as_users, create_mock_image, execsql, prepsql,
-    storage
+from cdedb.backend.common import affirm_validation as affirm
+from cdedb.common import (
+    PERSONA_CDE_FIELDS, PERSONA_EVENT_FIELDS, PERSONA_ML_FIELDS, ArchiveError,
+    CdEDBObject, PrivilegeError, RequestState, get_hash, merge_dicts, nearly_now, now,
 )
-
+from cdedb.validation import PERSONA_CDE_CREATION
+from tests.common import (
+    ANONYMOUS, USER_DICT, BackendTest, as_users, create_mock_image, execsql, prepsql,
+    storage,
+)
 
 PERSONA_TEMPLATE = {
     'username': "zelda@example.cde",
@@ -100,7 +100,7 @@ class TestCoreBackend(BackendTest):
 
             # Validate ml data
             persona = self.core.get_ml_user(self.key, persona_id)
-            affirm(vtypes.Persona, persona, _ignore_warnings=True)
+            affirm(vtypes.Persona, persona)
 
             # Validate event data if applicable
             if not persona['is_event_realm']:
@@ -108,10 +108,10 @@ class TestCoreBackend(BackendTest):
 
             persona = self.core.get_event_user(self.key, persona_id)
             if persona_id != USER_DICT["inga"]["id"]:
-                affirm(vtypes.Persona, persona, _ignore_warnings=True)
+                affirm(vtypes.Persona, persona)
             else:
                 with self.assertRaises(ValueError) as cm:
-                    affirm(vtypes.Persona, persona, _ignore_warnings=True)
+                    affirm(vtypes.Persona, persona)
                     self.assertIn("A birthday must be in the past. (birthday)",
                                   cm.exception.args)
 
@@ -121,10 +121,10 @@ class TestCoreBackend(BackendTest):
 
             persona = self.core.get_total_persona(self.key, persona_id)
             if persona_id != USER_DICT["inga"]["id"]:
-                affirm(vtypes.Persona, persona, _ignore_warnings=True)
+                affirm(vtypes.Persona, persona)
             else:
                 with self.assertRaises(ValueError) as cm:
-                    affirm(vtypes.Persona, persona, _ignore_warnings=True)
+                    affirm(vtypes.Persona, persona)
                     self.assertIn("A birthday must be in the past. (birthday)",
                                   cm.exception.args)
 
@@ -268,6 +268,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
         })
         self.assertEqual(data, new_data)
         expectation = {
@@ -311,6 +312,7 @@ class TestCoreBackend(BackendTest):
                 'is_ml_admin': False,
                 'is_ml_realm': False,
                 'is_cdelokal_admin': False,
+                'is_auditor': False,
                 'is_purged': False,
                 'is_searchable': False,
                 'location': None,
@@ -387,6 +389,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
         })
         self.assertEqual(data, new_data)
 
@@ -424,6 +427,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
         })
         self.assertEqual(data, new_data)
 
@@ -448,6 +452,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
         })
         self.assertEqual(data, new_data)
 
@@ -486,6 +491,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
         })
         self.assertEqual(data, new_data)
 
@@ -509,14 +515,15 @@ class TestCoreBackend(BackendTest):
                 if persona[key] is None:
                     persona[key] = True
         merge_dicts(data, persona)
-        self.assertLess(0, self.core.change_persona_realms(self.key, data))
+        change_note = "Bereichsänderung"
+        self.assertLess(0, self.core.change_persona_realms(self.key, data, change_note))
         log_entry = {
             'id': 1001,
             'ctime': nearly_now(),
             'code': const.CoreLogCodes.realm_change,
             'persona_id': persona_id,
             'submitted_by': self.user['id'],
-            'change_note': 'Bereiche geändert.'
+            'change_note': change_note,
         }
         _, expected_log = self.core.retrieve_log(self.key)
         self.assertIn(log_entry, expected_log)
@@ -653,6 +660,8 @@ class TestCoreBackend(BackendTest):
             'reviewer': None,
             'attachment_hash': None,
             'birth_name': None,
+            'pevent_id': None,
+            'pcourse_id': None,
         })
         value = self.core.genesis_get_case(self.key, case_id)
         del value['ctime']
@@ -685,6 +694,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
             'id': new_id,
             'display_name': 'Zelda',
             'is_active': True,
@@ -729,6 +739,8 @@ class TestCoreBackend(BackendTest):
             'telephone': None,
             'attachment_hash': None,
             'birth_name': None,
+            'pevent_id': None,
+            'pcourse_id': None,
         })
         value = self.core.genesis_get_case(self.key, case_id)
         del value['ctime']
@@ -760,6 +772,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
             'id': new_id,
             'display_name': 'Zelda',
             'is_active': True,
@@ -793,6 +806,8 @@ class TestCoreBackend(BackendTest):
             'location': "Marcuria",
             'country': "AQ",
             'attachment_hash': attachment_hash,
+            'pevent_id': None,
+            'pcourse_id': None,
         }
         self.assertFalse(self.core.genesis_attachment_usage(
             self.key, attachment_hash))
@@ -823,8 +838,7 @@ class TestCoreBackend(BackendTest):
         expectation.update(update)
         new_id = self.core.genesis(self.key, case_id)
         self.assertLess(0, new_id)
-        expectation = {k: v for k, v in expectation.items() if
-                       k in PERSONA_CDE_FIELDS}
+        expectation = {k: v for k, v in expectation.items() if k in PERSONA_CDE_FIELDS}
         expectation.update({
             'is_meta_admin': False,
             'is_archived': False,
@@ -837,6 +851,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_purged': False,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
             'id': new_id,
             'display_name': 'Zelda',
             'is_active': True,
@@ -947,6 +962,7 @@ class TestCoreBackend(BackendTest):
             'is_ml_admin': False,
             'is_ml_realm': True,
             'is_cdelokal_admin': False,
+            'is_auditor': False,
             'is_purged': False,
             'is_searchable': True,
             'username': 'berta@example.cde'}
@@ -960,10 +976,10 @@ class TestCoreBackend(BackendTest):
             'country': "DE",
             'gender': 1,
             'location': 'Utopia',
-            'mobile': '0163/123456789',
+            'mobile': '+4916312345678',
             'name_supplement': 'MdB',
             'postal_code': '34576',
-            'telephone': '+49 (5432) 987654321',
+            'telephone': '+495432987654321',
             'title': 'Dr.',
             })
         self.assertEqual(expectation, self.core.get_event_user(self.key, 2))
@@ -984,7 +1000,7 @@ class TestCoreBackend(BackendTest):
             'paper_expuls': True,
             'postal_code2': '8XA 45-$',
             'specialisation': 'Alles\nUnd noch mehr',
-            'telephone': '+49 (5432) 987654321',
+            'telephone': '+495432987654321',
             'timeline': 'Überall',
             'trial_member': False,
             'username': 'berta@example.cde',
@@ -1065,11 +1081,8 @@ class TestCoreBackend(BackendTest):
         self.core.archive_persona(self.key, 4, "Archived for testing.")
         self.core.dearchive_persona(self.key, 4, new_username="daniel@example.cde")
         # The following call sometimes failed with the error "editing
-        # archived members impossbile". The solution may be to add some
-        # sleep to let the DB settle, but this seems kind of bogus.
-        #
-        # import time
-        # time.sleep(1)
+        # archived members impossbile". This could be pampered over with a
+        # sleep to let the DB settle, but this would only hide a real problem.
         data = {
             'id': 4,
             'is_active': True,
@@ -1275,21 +1288,27 @@ class TestCoreBackend(BackendTest):
 
     @as_users("janis")
     def test_list_personas(self) -> None:
-        reality = self.core.list_all_personas(self.key, is_active=True)
-        active_personas = {1, 2, 3, 4, 5, 6, 7, 9, 10, 11, 13, 14, 16, 17, 18, 22,
-                           23, 27, 32, 48, 100}
-        self.assertEqual(active_personas, reality)
+        all_personas = tuple(e for e in self.get_sample_data("core.personas").values()
+                             if not e['is_archived'])
         reality = self.core.list_all_personas(self.key, is_active=False)
-        self.assertEqual(active_personas | {15}, reality)
+        self.assertEqual(reality, {e['id'] for e in all_personas})
+        active_personas = {e['id'] for e in all_personas if e['is_active']}
+        reality = self.core.list_all_personas(self.key, is_active=True)
+        self.assertEqual(reality, active_personas)
+        current_active_members = {
+            e['id'] for e in all_personas if e['is_member'] and e['is_active']}
         reality = self.core.list_current_members(self.key, is_active=True)
-        self.assertEqual({1, 2, 3, 6, 7, 9, 100}, reality)
+        self.assertEqual(reality, current_active_members)
+        current_members = {e['id'] for e in all_personas if e['is_member']}
         reality = self.core.list_current_members(self.key, is_active=False)
-        self.assertEqual({1, 2, 3, 6, 7, 9, 15, 100}, reality)
+        self.assertEqual(reality, current_members)
+        all_moderators = {
+            e['persona_id'] for e in self.get_sample_data("ml.moderators").values()}
         reality = self.core.list_all_moderators(self.key)
-        self.assertEqual({1, 2, 3, 4, 5, 7, 9, 10, 11, 15, 23, 27, 100}, reality)
+        self.assertEqual(reality, all_moderators)
         MT = const.MailinglistTypes
-        reality = self.core.list_all_moderators(self.key, {MT.member_moderated_opt_in,
-                                                           MT.cdelokal})
+        reality = self.core.list_all_moderators(
+            self.key, {MT.member_moderated_opt_in, MT.cdelokal})
         self.assertEqual({2, 5, 9, 100}, reality)
 
     @as_users("vera")
@@ -1320,7 +1339,7 @@ class TestCoreBackend(BackendTest):
              'change_note': None,
              'code': const.CoreLogCodes.persona_creation,
              'ctime': nearly_now(),
-             'persona_id': 1001,
+             'persona_id': new_persona_id,
              'submitted_by': self.user['id']},
             {'id': 1002,
              'change_note': 'zeldax@example.cde',
@@ -1347,8 +1366,29 @@ class TestCoreBackend(BackendTest):
     @as_users("vera")
     def test_changelog_meta(self) -> None:
         expectation = self.get_sample_data(
-            "core.changelog", range(1, 32),
-            ("id", "submitted_by", "reviewed_by", "ctime", "generation",
-             "change_note", "code", "persona_id"))
+            "core.changelog", ids=None,
+            keys=("id", "submitted_by", "reviewed_by", "ctime", "generation",
+                  "change_note", "code", "persona_id"))
         self.assertEqual((len(expectation), tuple(expectation.values())),
                          self.core.retrieve_changelog_meta(self.key))
+
+    @as_users("katarina")
+    def test_auditor(self) -> None:
+        for retriever, table in (
+            (self.core.retrieve_log, "core.log"),
+            (self.core.retrieve_changelog_meta, "core.changelog"),
+            (self.cde.retrieve_cde_log, "cde.log"),
+            (self.cde.retrieve_finance_log, "cde.finance_log"),
+            (self.event.retrieve_log, "event.log"),
+            (self.assembly.retrieve_log, "assembly.log"),
+            (self.ml.retrieve_log, "ml.log"),
+        ):
+            with self.subTest(log=table):
+                keys = None
+                if table == "core.changelog":
+                    keys = ("change_note", "code", "ctime", "generation", "id",
+                            "persona_id", "reviewed_by", "submitted_by")
+                self.assertLogEqual(
+                    tuple(self.get_sample_data(table, keys=keys).values()),
+                    log_retriever=retriever,  # type: ignore[arg-type]
+                )
