@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import collections.abc
 import copy
@@ -7,14 +8,15 @@ import decimal
 import json
 from typing import Any, Dict, List
 
+import freezegun
 import psycopg2
 import pytz
 
 import cdedb.database.constants as const
 from cdedb.backend.common import cast_fields
 from cdedb.common import (
-    CdEDBObject, CdEDBObjectMap, InfiniteEnum, PartialImportError, PrivilegeError,
-    CourseFilterPositions, nearly_now
+    CdEDBObject, CdEDBObjectMap, CourseFilterPositions, InfiniteEnum,
+    PartialImportError, PrivilegeError, nearly_now, now,
 )
 from cdedb.query import Query, QueryOperators, QueryScope
 from tests.common import USER_DICT, BackendTest, as_users, json_keys_to_int, storage
@@ -74,6 +76,8 @@ class TestEventBackend(BackendTest):
                              'min_choices': 3,
                              'sortkey': 1}
                     },
+                    'fee_modifiers': {
+                    },
                     'title': "First coming",
                     'shortname': "first",
                     'part_begin': datetime.date(2109, 8, 7),
@@ -89,6 +93,14 @@ class TestEventBackend(BackendTest):
                              'min_choices': 1,
                              'sortkey': 1}
                     },
+                    'fee_modifiers': {
+                        -1: {
+                            'amount': decimal.Decimal("-7.00"),
+                            # TODO allow specifying a negative id here?
+                            'field_id': 1003,
+                            'modifier_name': "is_child",
+                        }
+                    },
                     'title': "Second coming",
                     'shortname': "second",
                     'part_begin': datetime.date(2110, 8, 7),
@@ -101,31 +113,26 @@ class TestEventBackend(BackendTest):
                 -1: {
                     'association': 1,
                     'field_name': "instrument",
-                    'kind': 1,
+                    'kind': const.FieldDatatypes.str,
                     'entries': None,
+                    'checkin': False,
                 },
                 -2: {
                     'association': 1,
                     'field_name': "preferred_excursion_date",
-                    'kind': 5,
+                    'kind': const.FieldDatatypes.date,
                     'entries': [["2109-08-16", "In the first coming"],
                                 ["2110-08-16", "During the second coming"]],
+                    'checkin': True,
                 },
                 -3: {
                     'association': const.FieldAssociations.registration,
                     'field_name': "is_child",
                     'kind': const.FieldDatatypes.bool,
                     'entries': None,
+                    'checkin': False,
                 }
             },
-            'fee_modifiers': {
-                -1: {
-                    'amount': decimal.Decimal("-7.00"),
-                    'field_id': 1003,  # TODO allow specifying a negative id here?
-                    'modifier_name': "is_child",
-                    'part_id': 1002,  # TODO allow specifying a negative id here?
-                }
-            }
         }
         new_id = self.event.create_event(self.key, data)
         # back to normal mode
@@ -147,15 +154,19 @@ class TestEventBackend(BackendTest):
         data['end'] = datetime.date(2110, 8, 20)
         data['is_open'] = True
         # TODO dynamically adapt ids from the database result
+        data['parts'][-1].update({'id': 1001})
+        data['parts'][-2].update({'id': 1002})
         data['parts'][-1]['tracks'][-1].update({'id': 1001, 'part_id': 1001})
         data['parts'][-2]['tracks'][-1].update({'id': 1002, 'part_id': 1002})
         data['tracks'] = {1001: data['parts'][-1]['tracks'][-1],
                           1002: data['parts'][-2]['tracks'][-1]}
+        data['parts'][-2]['fee_modifiers'][-1].update({'id': 1001, 'part_id': 1002})
+        data['fee_modifiers'] = {1001: data['parts'][-2]['fee_modifiers'][-1]}
         # correct part and field ids
         tmp = self.event.get_event(self.key, new_id)
         part_map = {}
         for part in tmp['parts']:
-            for oldpart in data['parts'].keys():
+            for oldpart in data['parts']:
                 if tmp['parts'][part]['title'] == data['parts'][oldpart]['title']:
                     part_map[tmp['parts'][part]['title']] = part
                     data['parts'][part] = data['parts'][oldpart]
@@ -165,11 +176,18 @@ class TestEventBackend(BackendTest):
                         set(x['title'] for x in data['parts'][part]['tracks'].values()),
                         set(x['title'] for x in tmp['parts'][part]['tracks'].values()))
                     data['parts'][part]['tracks'] = tmp['parts'][part]['tracks']
+                    self.assertEqual(
+                        set(x['modifier_name']
+                            for x in data['parts'][part]['fee_modifiers'].values()),
+                        set(x['modifier_name']
+                            for x in tmp['parts'][part]['fee_modifiers'].values()))
+                    data['parts'][part]['fee_modifiers'] = (
+                        tmp['parts'][part]['fee_modifiers'])
                     del data['parts'][oldpart]
                     break
         field_map = {}
         for field in tmp['fields']:
-            for oldfield in data['fields'].keys():
+            for oldfield in data['fields']:
                 if (tmp['fields'][field]['field_name']
                         == data['fields'][oldfield]['field_name']):
                     field_map[tmp['fields'][field]['field_name']] = field
@@ -177,16 +195,6 @@ class TestEventBackend(BackendTest):
                     data['fields'][field]['id'] = field
                     data['fields'][field]['event_id'] = new_id
                     del data['fields'][oldfield]
-                    break
-        fee_modifier_map = {}
-        for mod in tmp['fee_modifiers']:
-            for oldmod in data['fee_modifiers'].keys():
-                if (tmp['fee_modifiers'][mod]['modifier_name']
-                        == data['fee_modifiers'][oldmod]['modifier_name']):
-                    fee_modifier_map[tmp['fee_modifiers'][mod]['modifier_name']] = mod
-                    data['fee_modifiers'][mod] = data['fee_modifiers'][oldmod]
-                    data['fee_modifiers'][mod]['id'] = mod
-                    del data['fee_modifiers'][oldmod]
                     break
 
         self.assertEqual(data,
@@ -214,25 +222,37 @@ class TestEventBackend(BackendTest):
             'fee': decimal.Decimal("1.23"),
             'waitlist_field': None,
             'tracks': {
-                1002: {'title': "Second lecture v2",
-                       'shortname': "Second v2",
-                       'num_choices': 5,
-                       'min_choices': 4,
-                       'sortkey': 3}
-            }}
+                1002: {
+                    'title': "Second lecture v2",
+                    'shortname': "Second v2",
+                    'num_choices': 5,
+                    'min_choices': 4,
+                    'sortkey': 3,
+                }
+            },
+            'fee_modifiers': {
+                1001: {
+                    'modifier_name': "ist_kind",
+                    'amount': decimal.Decimal("3.33"),
+                    'field_id': 1003,
+                },
+            },
+        }
         newfield = {
-            'association': 3,
+            'association': const.FieldAssociations.lodgement,
             'field_name': "kuea",
-            'kind': 1,
+            'kind': const.FieldDatatypes.str,
             'entries': None,
+            'checkin': False,
         }
         changed_field = {
-            'association': 2,
-            'kind': 5,
+            'association': const.FieldAssociations.course,
+            'kind': const.FieldDatatypes.date,
             'entries': [
                 ["2110-08-15", "early second coming"],
                 ["2110-08-17", "late second coming"],
             ],
+            'checkin': True,
         }
         self.event.set_event(self.key, {
             'id': new_id,
@@ -256,16 +276,20 @@ class TestEventBackend(BackendTest):
                 data['parts'][part] = newpart
                 data['parts'][part]['id'] = part
                 data['parts'][part]['event_id'] = new_id
-                self.assertEqual(set(x['title']
-                                     for x in data['parts'][part]['tracks'].values()),
-                                 set(x['title']
-                                     for x in tmp['parts'][part]['tracks'].values()))
+                self.assertEqual(
+                    set(x['title'] for x in data['parts'][part]['tracks'].values()),
+                    set(x['title'] for x in tmp['parts'][part]['tracks'].values()))
+                for track in tmp['parts'][part]['tracks']:
+                    tmp['parts'][part]['tracks'][track]['id'] = track
                 data['parts'][part]['tracks'] = tmp['parts'][part]['tracks']
+                data['parts'][part]['fee_modifiers'] = (
+                    tmp['parts'][part]['fee_modifiers'])
         del data['parts'][part_map["First coming"]]
         changed_part['id'] = part_map["Second coming"]
         changed_part['event_id'] = new_id
         changed_part['shortname'] = "second"
         changed_part['tracks'][1002].update({'part_id': 1002, 'id': 1002})
+        changed_part['fee_modifiers'][1001].update({'part_id': 1002, 'id': 1001})
         data['parts'][part_map["Second coming"]] = changed_part
         for field in tmp['fields']:
             if tmp['fields'][field]['field_name'] == "kuea":
@@ -281,23 +305,29 @@ class TestEventBackend(BackendTest):
         data['begin'] = datetime.date(2110, 9, 8)
         data['end'] = datetime.date(2111, 8, 20)
         # TODO dynamically adapt ids from the database result
-        data['tracks'] = {1002: {'id': 1002,
-                                 'part_id': 1002,
-                                 'title': 'Second lecture v2',
-                                 'shortname': "Second v2",
-                                 'num_choices': 5,
-                                 'min_choices': 4,
-                                 'sortkey': 3},
-                          1003: {'id': 1003,
-                                 'part_id': 1003,
-                                 'title': 'Third lecture',
-                                 'shortname': 'Third',
-                                 'num_choices': 2,
-                                 'min_choices': 2,
-                                 'sortkey': 2}}
+        data['tracks'] = {
+            1002: {
+                'id': 1002,
+                'part_id': 1002,
+                'title': 'Second lecture v2',
+                'shortname': "Second v2",
+                'num_choices': 5,
+                'min_choices': 4,
+                'sortkey': 3,
+            },
+            1003: {
+                'id': 1003,
+                'part_id': 1003,
+                'title': 'Third lecture',
+                'shortname': 'Third',
+                'num_choices': 2,
+                'min_choices': 2,
+                'sortkey': 2,
+            },
+        }
+        data['fee_modifiers'] = changed_part['fee_modifiers']
 
-        self.assertEqual(data,
-                         self.event.get_event(self.key, new_id))
+        self.assertEqual(data, self.event.get_event(self.key, new_id))
 
         self.assertNotIn(new_id, old_events)
         new_events = self.event.list_events(self.key)
@@ -563,7 +593,8 @@ class TestEventBackend(BackendTest):
         }
         with self.assertRaises(ValueError) as cm:
             self.event.set_questionnaire(self.key, event_id, questionnaire)
-        self.assertIn("Must not duplicate field. (field_id)", cm.exception.args)
+        self.assertEqual("Must not duplicate field ('brings_balls'). (field_id)",
+                         cm.exception.args[0] % cm.exception.args[1])
 
         # Event mustn't have registrations to alter fee modifiers.
         reg_ids = self.event.list_registrations(self.key, event_id)
@@ -575,28 +606,34 @@ class TestEventBackend(BackendTest):
         self.event.set_questionnaire(self.key, event_id, None)
         data = {
             'id': event_id,
-            'fee_modifiers':
-                {
-                    anid: None
-                    for anid in old_event['fee_modifiers']
+            'parts': {
+                part_id: {
+                    'fee_modifiers': {
+                        modifier_id: None
+                        for modifier_id in old_event['parts'][part_id]['fee_modifiers']
+                    }
                 }
+                for part_id in old_event['parts']
+            },
         }
         self.event.set_event(self.key, data)
 
         data = {
             'id': event_id,
-            'fee_modifiers': {
-                -1: {
-                    'field_id': 7,
-                    'modifier_name': "is_child",
-                    'amount': decimal.Decimal("-8.00"),
-                    'part_id': list(old_event['parts'])[0],
-                },
-                -2: {
-                    'field_id': 7,
-                    'modifier_name': "is_child2",
-                    'amount': decimal.Decimal("-7.00"),
-                    'part_id': list(old_event['parts'])[0],
+            'parts': {
+                list(old_event['parts'])[0]: {
+                    'fee_modifiers': {
+                        -1: {
+                            'field_id': 7,
+                            'modifier_name': "is_child",
+                            'amount': decimal.Decimal("-8.00"),
+                        },
+                        -2: {
+                            'field_id': 7,
+                            'modifier_name': "is_child2",
+                            'amount': decimal.Decimal("-7.00"),
+                        },
+                    },
                 },
             },
         }
@@ -633,7 +670,8 @@ class TestEventBackend(BackendTest):
         expectation = {
             'anzahl_GROSSBUCHSTABEN': 4,
             'arrival': datetime.datetime(2222, 11, 9, 8, 55, 44, tzinfo=pytz.utc),
-            'lodge': 'Die üblichen Verdächtigen :)',
+            'lodge': 'Die üblichen Verdächtigen, insb. Berta Beispiel und '
+                     'garcia@example.cde :)',
             'is_child': False,
         }
         self.assertEqual(expectation, data['fields'])
@@ -990,7 +1028,8 @@ class TestEventBackend(BackendTest):
                 'event_id': 1,
                 'fields': {
                     'anzahl_GROSSBUCHSTABEN': 4,
-                    'lodge': 'Die üblichen Verdächtigen :)',
+                    'lodge': 'Die üblichen Verdächtigen, insb. Berta Beispiel '
+                             'und garcia@example.cde :)',
                     'is_child': False,
                 },
                 'list_consent': True,
@@ -1452,7 +1491,7 @@ class TestEventBackend(BackendTest):
                      'default_value': None,
                      'info': None,
                      'pos': 0,
-                     'readonly': None,
+                     'readonly': False,
                      'input_size': None,
                      'title': 'Ich bin unter 13 Jahre alt.',
                      'kind': const.QuestionnaireUsages.registration,
@@ -1589,7 +1628,7 @@ class TestEventBackend(BackendTest):
                     'default_value': None,
                     'info': "Du kannst freiwillig etwas mehr bezahlen um zukünftige"
                             " Akademien zu unterstützen.",
-                    'readonly': None,
+                    'readonly': False,
                     'input_size': None,
                     'title': "Ich möchte den Solidaritätszuschlag bezahlen.",
                 },
@@ -2005,7 +2044,8 @@ class TestEventBackend(BackendTest):
             if name != "Test-Query":
                 self.assertIn(name, expectation)
                 q = expectation[name]
-                self.assertEqual(set(q.fields_of_interest), set(query.fields_of_interest))
+                self.assertEqual(set(q.fields_of_interest),
+                                 set(query.fields_of_interest))
                 self.assertEqual(set(q.constraints), set(query.constraints))
                 self.assertEqual(set(q.order), set(query.order))
                 self.assertEqual(q.query_id, query.query_id)
@@ -2116,7 +2156,6 @@ class TestEventBackend(BackendTest):
         self.assertTrue(self.event.lock_event(self.key, 1))
         data = self.event.export_event(self.key, 1)
         new_data = copy.deepcopy(data)
-        del new_data['CDEDB_EXPORT_EVENT_VERSION']
         stored_data = copy.deepcopy(data)
         # Apply some changes
 
@@ -2297,7 +2336,8 @@ class TestEventBackend(BackendTest):
                 'event_id': 1,
                 'field_name': 'behaviour',
                 'id': 11000,
-                'kind': 1,
+                'kind': const.FieldDatatypes.str,
+                'checkin': False,
             },
             11001: {
                 'association': const.FieldAssociations.registration,
@@ -2306,6 +2346,7 @@ class TestEventBackend(BackendTest):
                 'field_name': "solidarity",
                 'id': 11001,
                 'kind': const.FieldDatatypes.bool,
+                'checkin': False,
             }
         })
         # questionnaire rows
@@ -2507,6 +2548,7 @@ class TestEventBackend(BackendTest):
                 'field_name': 'behaviour',
                 'id': 1001,
                 'kind': const.FieldDatatypes.str,
+                'checkin': False,
             },
             1002: {
                 'association': const.FieldAssociations.registration,
@@ -2515,6 +2557,7 @@ class TestEventBackend(BackendTest):
                 'field_name': 'solidarity',
                 'id': 1002,
                 'kind': const.FieldDatatypes.bool,
+                'checkin': False,
             },
         })
         stored_data['event.fee_modifiers'][1001] = {
@@ -3076,24 +3119,30 @@ class TestEventBackend(BackendTest):
 
         data = {
             'id': event_id,
-            'fee_modifiers': {
-                -1: {
-                    'amount': decimal.Decimal("1.00"),
-                    'field_id': 1001,
-                    'modifier_name': unique_name,
-                    'part_id': part_id,
+            'parts': {
+                part_id: {
+                    'fee_modifiers': {
+                        -1: {
+                            'amount': decimal.Decimal("1.00"),
+                            'field_id': 1001,
+                            'modifier_name': unique_name,
+                        },
+                    },
                 },
             },
         }
         self.event.set_event(self.key, data)
         data = {
             'id': event_id,
-            'fee_modifiers': {
-                -1: {
-                    'amount': decimal.Decimal("1.00"),
-                    'field_id': 1001,
-                    'modifier_name': unique_name + "2",
-                    'part_id': part_id,
+            'parts': {
+                part_id: {
+                    'fee_modifiers': {
+                        -1: {
+                            'amount': decimal.Decimal("1.00"),
+                            'field_id': 1001,
+                            'modifier_name': unique_name + "2",
+                        },
+                    },
                 },
             },
         }
@@ -3102,12 +3151,15 @@ class TestEventBackend(BackendTest):
             self.event.set_event(self.key, data)
         data = {
             'id': event_id,
-            'fee_modifiers': {
-                -1: {
-                    'amount': decimal.Decimal("1.00"),
-                    'field_id': 1003,
-                    'modifier_name': unique_name,
-                    'part_id': part_id,
+            'parts': {
+                part_id: {
+                    'fee_modifiers': {
+                        -1: {
+                            'amount': decimal.Decimal("1.00"),
+                            'field_id': 1003,
+                            'modifier_name': unique_name,
+                        },
+                    },
                 },
             },
         }
@@ -3116,12 +3168,15 @@ class TestEventBackend(BackendTest):
             self.event.set_event(self.key, data)
         data = {
             'id': event_id,
-            'fee_modifiers': {
-                -1: {
-                    'amount': decimal.Decimal("1.00"),
-                    'field_id': 1003,
-                    'modifier_name': unique_name + "2",
-                    'part_id': part_id,
+            'parts': {
+                part_id: {
+                    'fee_modifiers': {
+                        -1: {
+                            'amount': decimal.Decimal("1.00"),
+                            'field_id': 1003,
+                            'modifier_name': unique_name + "2",
+                        },
+                    },
                 },
             },
         }
@@ -3158,26 +3213,30 @@ class TestEventBackend(BackendTest):
         field_links = (
             (1001, None, None),
             (1001, psycopg2.IntegrityError, None),
-            (1002, ValueError, "Fee Modifier linked to non-bool field."),
-            (1003, ValueError, "Fee Modifier linked to non-registration field."),
+            (1002, ValueError, "Unfit field for fee_modifier."),
+            (1003, ValueError, "Unfit field for fee_modifier."),
         )
         for field_id, error, error_msg in field_links:
             data = {
                 'id': event_id,
-                'fee_modifiers': {
-                    -1: {
-                        'modifier_name': 'solidarity',
-                        'amount': decimal.Decimal("-12.50"),
-                        'field_id': field_id,
-                        'part_id': list(event["parts"])[0],
-                    }
-                }
+                'parts': {
+                    list(event['parts'])[0]: {
+                        'fee_modifiers': {
+                            -1: {
+                                'modifier_name': 'solidarity',
+                                'amount': decimal.Decimal("-12.50"),
+                                'field_id': field_id,
+                            }
+                        },
+                    },
+                },
             }
             if error:
                 with self.assertRaises(error) as cm:
                     self.event.set_event(self.key, data)
                 if error_msg is not None:
-                    self.assertEqual(error_msg, cm.exception.args[0])
+                    self.assertEqual(error_msg,
+                                     cm.exception.args[0] % cm.exception.args[1])
             else:
                 self.assertTrue(self.event.set_event(self.key, data))
         reg_data = {
@@ -3247,33 +3306,55 @@ class TestEventBackend(BackendTest):
                         'status': const.RegistrationPartStati.waitlist,
                     },
                     2: {
-                        'status': const.RegistrationPartStati.waitlist
-                    } if anid in {2, 3} else {},
+                        'status': (const.RegistrationPartStati.waitlist
+                                   if anid in {2, 3}
+                                   else const.RegistrationPartStati.participant)
+                    },
                     3: {
-                        'status': const.RegistrationPartStati.waitlist
-                    } if anid in {2, 3} else {},
+                        'status': (const.RegistrationPartStati.waitlist
+                                   if anid in {2, 3}
+                                   else const.RegistrationPartStati.participant)
+                    },
                 },
                 'fields': {
-                    'waitlist': i,
+                    'waitlist': i+1,
                 },
             }
             for i, anid in enumerate((5, 4, 3, 2, 1))
         ]
         for rdata in regs:
             self.event.set_registration(self.key, rdata)
-        self.assertEqual({1: [5, 4, 3, 2, 1], 2: [3, 2], 3: [3, 2]},
-                         self.event.get_waitlist(self.key, event_id=1))
+        # Registration 3 belongs to Garcia (persona_id 7).
+        expectation = {1: [5, 4, 3, 2, 1], 2: [3, 2], 3: [3, 2]}
+        self.assertEqual(expectation, self.event.get_waitlist(self.key, event_id=1))
         self.assertEqual({1: 3, 2: 1, 3: 1},
                          self.event.get_waitlist_position(self.key, event_id=1))
+        # Registration 2 belongs to Emilia (persona_id 5).
         self.assertEqual({1: 4, 2: 2, 3: 2},
                          self.event.get_waitlist_position(
                              self.key, event_id=1, persona_id=5))
+        # Unset waitlist field data.
+        reg_id = 4
+        reg_data = {
+            'id': reg_id,
+            'fields': {
+                'waitlist': None,
+            },
+        }
+        self.event.set_registration(self.key, reg_data)
+        # The altered registration will be placed first in the waitlist, because
+        # it defaults to 0.
+        for waitlist in expectation.values():
+            if reg_id in waitlist:
+                waitlist.remove(reg_id)
+                waitlist.insert(0, reg_id)
+        self.assertEqual(expectation, self.event.get_waitlist(self.key, event_id=1))
+
+        # Check that users can check their own waitlist position.
         self.login(USER_DICT["emilia"])
-        self.event._get_waitlist(self.key, event_id=1)
         self.assertEqual({1: 4, 2: 2, 3: 2},
-                         self.event.get_waitlist_position(
-                             self.key, event_id=1))
-        with self.assertRaises(PrivilegeError) as cm:
+                         self.event.get_waitlist_position(self.key, event_id=1))
+        with self.assertRaises(PrivilegeError):
             self.event.get_waitlist_position(
                 self.key, event_id=1, persona_id=1)
 
@@ -3396,6 +3477,7 @@ class TestEventBackend(BackendTest):
                     'field_name': "instrument",
                     'kind': 1,
                     'entries': None,
+                    'checkin': False,
                 },
                 -2: {
                     'association': 1,
@@ -3403,6 +3485,7 @@ class TestEventBackend(BackendTest):
                     'kind': 5,
                     'entries': [["2109-8-16", "In the first coming"],
                                 ["2110-8-16", "During the second coming"]],
+                    'checkin': True,
                 },
             },
         }
@@ -3411,7 +3494,7 @@ class TestEventBackend(BackendTest):
         tmp = self.event.get_event(self.key, new_id)
         part_map = {}
         for part in tmp['parts']:
-            for oldpart in data['parts'].keys():
+            for oldpart in data['parts']:
                 if tmp['parts'][part]['title'] == data['parts'][oldpart]['title']:
                     part_map[tmp['parts'][part]['title']] = part
                     data['parts'][part] = data['parts'][oldpart]
@@ -3421,7 +3504,7 @@ class TestEventBackend(BackendTest):
                     break
         field_map = {}
         for field in tmp['fields']:
-            for oldfield in data['fields'].keys():
+            for oldfield in data['fields']:
                 if (tmp['fields'][field]['field_name']
                         == data['fields'][oldfield]['field_name']):
                     field_map[tmp['fields'][field]['field_name']] = field
@@ -3452,25 +3535,30 @@ class TestEventBackend(BackendTest):
             'part_end': datetime.date(2110, 9, 21),
             'fee': decimal.Decimal("1.23"),
             'tracks': {
-                1002: {'title': "Second lecture v2",  # hardcoded id 5
-                       'shortname': "Second v2",
-                       'num_choices': 5,
-                       'min_choices': 4,
-                       'sortkey': 3}}
+                1002: {
+                    'title': "Second lecture v2",  # hardcoded id 5
+                    'shortname': "Second v2",
+                    'num_choices': 5,
+                    'min_choices': 4,
+                    'sortkey': 3,
+                },
+            },
         }
         newfield = {
-            'association': 1,
+            'association': const.FieldAssociations.registration,
             'field_name': "kuea",
-            'kind': 1,
+            'kind': const.FieldDatatypes.date,
             'entries': None,
+            'checkin': False,
         }
         changed_field = {
-            'association': 1,
-            'kind': 5,
+            'association': const.FieldAssociations.registration,
+            'kind': const.FieldDatatypes.date,
             'entries': [
                 ["2110-8-15", "early second coming"],
                 ["2110-8-17", "late second coming"],
             ],
+            'checkin': True,
         }
         self.event.add_event_orgas(self.key, new_id, {2, 1})
         self.event.remove_event_orga(self.key, new_id, 2)
@@ -3644,217 +3732,217 @@ class TestEventBackend(BackendTest):
         expectation = (35, (
             {'id': 1001,
              'change_note': None,
-             'code': 1,
+             'code': const.EventLogCodes.event_created,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1002,
              'change_note': None,
-             'code': 10,
+             'code': const.EventLogCodes.orga_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': 2,
              'submitted_by': self.user['id']},
             {'id': 1003,
              'change_note': None,
-             'code': 10,
+             'code': const.EventLogCodes.orga_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': 7,
              'submitted_by': self.user['id']},
             {'id': 1004,
-             'change_note': 'First lecture',
-             'code': 35,
+             'change_note': 'instrument',
+             'code': const.EventLogCodes.field_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1005,
-             'change_note': 'First coming',
-             'code': 15,
+             'change_note': 'preferred_excursion_date',
+             'code': const.EventLogCodes.field_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1006,
-             'change_note': 'Second lecture',
-             'code': 35,
+             'change_note': 'First coming',
+             'code': const.EventLogCodes.part_created,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1007,
-             'change_note': 'Second coming',
-             'code': 15,
+             'change_note': 'First lecture',
+             'code': const.EventLogCodes.track_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1008,
-             'change_note': 'instrument',
-             'code': 20,
+             'change_note': 'Second coming',
+             'code': const.EventLogCodes.part_created,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1009,
-             'change_note': 'preferred_excursion_date',
-             'code': 20,
+             'change_note': 'Second lecture',
+             'code': const.EventLogCodes.track_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1010,
              'change_note': None,
-             'code': 10,
+             'code': const.EventLogCodes.orga_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': 1,
              'submitted_by': self.user['id']},
             {'id': 1011,
              'change_note': None,
-             'code': 11,
+             'code': const.EventLogCodes.orga_removed,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': 2,
              'submitted_by': self.user['id']},
             {'id': 1012,
              'change_note': None,
-             'code': 2,
+             'code': const.EventLogCodes.event_changed,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1013,
-             'change_note': 'Third lecture',
-             'code': 35,
+             'change_note': 'kuea',
+             'code': const.EventLogCodes.field_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1014,
-             'change_note': 'Third coming',
-             'code': 15,
+             'change_note': 'preferred_excursion_date',
+             'code': const.EventLogCodes.field_updated,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1015,
-             'change_note': 'Second lecture v2',
-             'code': 36,
+             'change_note': 'instrument',
+             'code': const.EventLogCodes.field_removed,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1016,
-             'change_note': 'Second coming',
-             'code': 16,
+             'change_note': 'Third coming',
+             'code': const.EventLogCodes.part_created,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1017,
-             'change_note': 'First lecture',
-             'code': 37,
+             'change_note': 'Third lecture',
+             'code': const.EventLogCodes.track_added,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1018,
-             'change_note': 'First coming',
-             'code': 17,
+             'change_note': 'Second coming',
+             'code': const.EventLogCodes.part_changed,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1019,
-             'change_note': 'kuea',
-             'code': 20,
+             'change_note': 'Second lecture v2',
+             'code': const.EventLogCodes.track_updated,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1020,
-             'change_note': 'preferred_excursion_date',
-             'code': 21,
+             'change_note': 'First lecture',
+             'code': const.EventLogCodes.track_removed,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1021,
-             'change_note': 'instrument',
-             'code': 22,
+             'change_note': 'First coming',
+             'code': const.EventLogCodes.part_deleted,
              'ctime': nearly_now(),
              'event_id': 1001,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1022,
              'change_note': 'Topos theory for the kindergarden',
-             'code': 42,
+             'code': const.EventLogCodes.course_segments_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1023,
              'change_note': 'Topos theory for the kindergarden',
-             'code': 40,
+             'code': const.EventLogCodes.course_created,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1024,
              'change_note': 'Topos theory for the kindergarden',
-             'code': 41,
+             'code': const.EventLogCodes.course_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1025,
              'change_note': 'Topos theory for the kindergarden',
-             'code': 42,
+             'code': const.EventLogCodes.course_segments_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1026,
              'change_note': None,
-             'code': 50,
+             'code': const.EventLogCodes.registration_created,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': 3,
              'submitted_by': self.user['id']},
             {'id': 1027,
              'change_note': "Boring change.",
-             'code': 51,
+             'code': const.EventLogCodes.registration_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': 9,
              'submitted_by': self.user['id']},
             {'id': 1028,
              'change_note': 'HY',
-             'code': 26,
+             'code': const.EventLogCodes.lodgement_created,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1029,
              'change_note': 'HY',
-             'code': 25,
+             'code': const.EventLogCodes.lodgement_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1030,
              'change_note': 'HY',
-             'code': 27,
+             'code': const.EventLogCodes.lodgement_deleted,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
              'submitted_by': self.user['id']},
             {'id': 1031,
              'change_note': None,
-             'code': 30,
+             'code': const.EventLogCodes.questionnaire_changed,
              'ctime': nearly_now(),
              'event_id': 1,
              'persona_id': None,
@@ -3862,3 +3950,41 @@ class TestEventBackend(BackendTest):
 
         result = self.event.retrieve_log(self.key, offset=offset)
         self.assertEqual(expectation, result)
+
+    def _create_registration(self, persona_id: int, event_id: int) -> int:
+        event = self.event.get_event(self.key, event_id)
+        return self.event.create_registration(self.key, {
+            'persona_id': persona_id,
+            'event_id': event['id'],
+            'mixed_lodging': True,
+            'list_consent': True,
+            'notes': None,
+            'parts': {
+                p_id: {'status': const.RegistrationPartStati.applied}
+                for p_id in event['parts']
+            },
+            'tracks': {
+                t_id: {}
+                for p_id in event['parts'] for t_id in event['parts'][p_id]['tracks']
+            }
+        })
+
+    @as_users("annika")
+    def test_registration_timestamps(self) -> None:
+        persona_id = self.user['id']
+        event_ids = [1, 2]
+        reg_ids = {}
+        base_time = now()
+        delta = datetime.timedelta(seconds=42)
+        with freezegun.freeze_time(base_time) as frozen_time:
+            for event_id in event_ids:
+                reg_id = self._create_registration(persona_id, event_id)
+                frozen_time.tick(delta)
+                self.event.set_registration(
+                    self.key, {'id': reg_id, 'notes': "Important change!"})
+                frozen_time.tick(delta)
+                reg_ids[event_id] = reg_id
+            for i, (event_id, reg_id) in enumerate(reg_ids.items()):
+                reg = self.event.get_registration(self.key, reg_id)
+                self.assertEqual(reg['ctime'], base_time + 2 * i * delta)
+                self.assertEqual(reg['mtime'], base_time + (2 * i + 1) * delta)
