@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import unittest
 
@@ -24,14 +25,16 @@ class TestDatabase(unittest.TestCase):
             self.assertIsInstance(conn, psycopg2.extensions.connection)
             self.assertIsInstance(conn, IrradiatedConnection)
         with self.assertRaises(ValueError):
-            _ = factory["nonexistentrole"]
+            # pylint: disable=pointless-statement
+            factory["nonexistentrole"]  # exception in __getitem__
 
     def test_less_users(self) -> None:
         factory = connection_pool_factory(
             _CONF["CDB_DATABASE_NAME"], ("cdb_anonymous", "cdb_admin"),
             _SECRECTSCONF, _CONF["DB_PORT"])
         with self.assertRaises(ValueError):
-            _ = factory["cdb_persona"]
+            # pylint: disable=pointless-statement
+            factory["cdb_persona"]  # exception in __getitem__
 
     def test_atomizer(self) -> None:
         factory = connection_pool_factory(
