@@ -1,12 +1,13 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import collections.abc
 import datetime
 import decimal
 import json
 import numbers
-from typing import Any, Dict, Set, Union, cast
 import unittest.mock
+from typing import Any, Dict, Set, Union, cast
 
 import cdedb.database.constants as const
 from cdedb.common import RequestState, now, xsorted
@@ -160,7 +161,7 @@ class TestCron(CronTest):
         ctime=(now() - datetime.timedelta(hours=6))))
     def test_genesis_remind_new(self) -> None:
         self.execute('genesis_remind')
-        self.assertEqual(["genesis_requests_pending"],
+        self.assertEqual(["genesis/genesis_requests_pending"],
                          [mail.template for mail in self.mails])
 
     @prepsql(genesis_template())
@@ -184,7 +185,7 @@ class TestCron(CronTest):
                         store={"tstamp": 1, "ids": [1001]}))
     def test_genesis_remind_older(self) -> None:
         self.execute('genesis_remind')
-        self.assertEqual(["genesis_requests_pending"],
+        self.assertEqual(["genesis/genesis_requests_pending"],
                          [mail.template for mail in self.mails])
 
     @storage
@@ -532,7 +533,7 @@ class TestCron(CronTest):
         self.assertEqual(
             mm_lists['witz'].subscribe.call_args_list,
             [umcall('new-anton@example.cde',
-                    display_name='Anton Armin A. Administrator',
+                    display_name='Anton Administrator',
                     pre_approved=True, pre_confirmed=True, pre_verified=True)])
         self.assertEqual(
             mm_lists['witz'].unsubscribe.call_args_list,
