@@ -1,18 +1,26 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import csv
 import re
 import unittest.mock
-from typing import Any, List, Tuple
+from typing import Any, List, Set, Tuple
+
+import webtest
 
 import cdedb.database.constants as const
-import cdedb.frontend.common
-import cdedb.ml_type_aux as ml_type
 from cdedb.common import ADMIN_VIEWS_COOKIE_NAME, CdEDBObject
-from cdedb.devsamples import MockHeldMessage, HELD_MESSAGE_SAMPLE
+from cdedb.devsamples import HELD_MESSAGE_SAMPLE, MockHeldMessage
 from cdedb.frontend.common import CustomCSVDialect
+from cdedb.ml_type_aux import CdeLokalMailinglist
 from cdedb.query import QueryOperators
 from tests.common import USER_DICT, FrontendTest, as_users, prepsql
+
+
+def _get_registration_part_stati(f: webtest.Form) -> Set[const.RegistrationPartStati]:
+    return set(filter(
+        None, (f.get("registration_stati", index=i).value
+               for i in range(len(const.RegistrationPartStati)))))
 
 
 class TestMlFrontend(FrontendTest):
@@ -147,14 +155,16 @@ class TestMlFrontend(FrontendTest):
         f['source_persona_id'] = USER_DICT['rowena']['DB-ID']
         f['target_persona_id'] = berta_id
         self.submit(f, check_notification=False)
-        msg = "Der Quellnutzer muss ein reiner Mailinglistennutzer und darf kein Admin sein."
+        msg = ("Der Quellnutzer muss ein reiner Mailinglistennutzer und darf kein Admin"
+               " sein.")
         self.assertValidationError('source_persona_id', msg)
 
         f = self.response.forms['merge-accounts']
         f['source_persona_id'] = USER_DICT['nina']['DB-ID']
         f['target_persona_id'] = berta_id
         self.submit(f, check_notification=False)
-        msg = "Der Quellnutzer muss ein reiner Mailinglistennutzer und darf kein Admin sein."
+        msg = ("Der Quellnutzer muss ein reiner Mailinglistennutzer und darf kein Admin"
+               " sein.")
         self.assertValidationError('source_persona_id', msg)
 
         f = self.response.forms['merge-accounts']
@@ -176,7 +186,8 @@ class TestMlFrontend(FrontendTest):
         f['source_persona_id'] = janis_id
         f['target_persona_id'] = berta_id
         self.submit(f, check_notification=False)
-        msg = "Beide Benutzer haben einen Bezug zu gleichen Mailinglisten: Witz des Tages"
+        msg = ("Beide Benutzer haben einen Bezug zu gleichen Mailinglisten: Witz des"
+               " Tages")
         self.assertPresence(msg, div='notifications')
 
         # ... so we resolve the blocking ...
@@ -442,6 +453,13 @@ class TestMlFrontend(FrontendTest):
         self.assertNonPresence("Versammlungslisten")
         self.assertNonPresence("Andere Mailinglisten")
 
+    @as_users("quintus")
+    def test_mailinglist_cde_admin(self) -> None:
+        self.traverse({'href': '/ml/$'},
+                      {'href': '/ml/mailinglist/list'},
+                      {'href': '/ml/mailinglist/7/show'},
+                      {'href': '/ml/mailinglist/7/management'})
+
     @as_users("nina", "berta")
     def test_mailinglist_management(self) -> None:
         self.traverse({'href': '/ml/$'},
@@ -449,7 +467,7 @@ class TestMlFrontend(FrontendTest):
                       {'href': '/ml/mailinglist/4/management'})
         self.assertTitle("Klatsch und Tratsch – Verwaltung")
         self.assertNonPresence("Inga Iota", div="moderator-list")
-        self.assertNonPresence("Anton Armin A. Administrator", div="moderator-list")
+        self.assertNonPresence("Anton Administrator", div="moderator-list")
         f = self.response.forms['addmoderatorform']
         # Check that you cannot add non-existing or archived moderators.
         errormsg = "Einige dieser Nutzer existieren nicht oder sind archiviert."
@@ -466,7 +484,7 @@ class TestMlFrontend(FrontendTest):
         self.submit(f)
         self.assertTitle("Klatsch und Tratsch – Verwaltung")
         self.assertPresence("Inga Iota", div="moderator-list")
-        self.assertPresence("Anton Armin A. Administrator",
+        self.assertPresence("Anton Administrator",
                             div="moderator-list")
         f = self.response.forms['removemoderatorform9']
         self.submit(f)
@@ -491,6 +509,31 @@ class TestMlFrontend(FrontendTest):
         self.submit(f)
         self.assertTitle("Klatsch und Tratsch – Erweiterte Verwaltung")
         self.assertNonPresence("zelda@example.cde")
+
+    @as_users("nina")
+    def test_mandatory_mailinglist(self) -> None:
+        self.traverse({'href': '/ml/$'},
+                      {'href': '/ml/mailinglist/1/'},
+                      {'href': '/ml/mailinglist/1/management'})
+        self.assertDivNotExists('removesubscriberform100')
+        self.traverse("Erweiterte Verwaltung")
+        self.assertTitle("Verkündungen – Erweiterte Verwaltung")
+        f = self.response.forms['addmodsubscriberform']
+        f['modsubscriber_ids'] = "DB-4-3"
+        self.submit(f)
+        self.assertTitle("Verkündungen – Erweiterte Verwaltung")
+        self.assertPresence("Daniel Dino")
+        f = self.response.forms['removemodsubscriberform4']
+        self.submit(f)
+        self.assertTitle("Verkündungen – Erweiterte Verwaltung")
+        self.assertNonPresence("Daniel Dino", div="modsubscriber-list")
+        self.traverse("Verwaltung")
+        self.assertPresence("Daniel Dino")
+        # Reload server- and client-side
+        self.ml.write_subscription_states(self.key, (1,))
+        self.traverse({'href': '/ml/mailinglist/1'},
+                      {'href': '/ml/mailinglist/1/management'})
+        self.assertNonPresence("Daniel")
 
     @as_users("nina", "berta")
     def test_advanced_management(self) -> None:
@@ -559,7 +602,7 @@ class TestMlFrontend(FrontendTest):
         f = self.response.forms['removemodunsubscriberform5']
         self.submit(f)
         self.assertTitle("Klatsch und Tratsch – Erweiterte Verwaltung")
-        self.assertNonPresence("Emilia E. Eventis", div="modunsubscriber-list")
+        self.assertNonPresence("Emilia", div="modunsubscriber-list")
         self.assertPresence("Emilia E. Eventis", div="unsubscriber-list")
 
         self.assertPresence("zelda@example.cde")
@@ -727,18 +770,18 @@ class TestMlFrontend(FrontendTest):
         self.traverse({'description': 'Mailingliste anlegen'})
         self.assertTitle("Mailingliste anlegen")
         f = self.response.forms['selectmltypeform']
-        f['ml_type'] = const.MailinglistTypes.member_mandatory.value
+        f['ml_type'] = const.MailinglistTypes.member_mandatory
         self.submit(f)
         f = self.response.forms['createlistform']
         self.assertEqual(f['maxsize'].value, '64')
         f['title'] = "Munkelwand"
-        f['mod_policy'] = 1
-        f['attachment_policy'] = 2
-        f['subject_prefix'] = "[munkel]"
+        f['mod_policy'] = const.ModerationPolicy.unmoderated
+        f['attachment_policy'] = const.AttachmentPolicy.pdf_only
+        f['subject_prefix'] = "munkel"
         f['maxsize'] = 512
         f['is_active'].checked = True
         f['notes'] = "Noch mehr Gemunkel."
-        f['domain'] = 1
+        f['domain'] = const.MailinglistDomain.lists
         f['local_part'] = 'munkelwand'
         # Check that there must be some moderators
         errormsg = "Darf nicht leer sein."
@@ -766,7 +809,7 @@ class TestMlFrontend(FrontendTest):
         self.submit(f)
         self.assertTitle("Munkelwand")
         self.assertPresence("Clown")
-        self.assertPresence("Garcia G. Generalis")
+        self.assertPresence("Garcia Generalis")
 
     @as_users("nina")
     def test_change_mailinglist(self) -> None:
@@ -779,7 +822,7 @@ class TestMlFrontend(FrontendTest):
         # Check if the attachment policy hint works
         self.assertPresence("Admin-Team")
         f = self.response.forms['changelistform']
-        f['domain'] = const.MailinglistDomain.testmail.value
+        f['domain'] = const.MailinglistDomain.testmail
         f['maxsize'] = "intentionally no valid maxsize"
         self.submit(f, check_notification=False)
         self.assertValidationError("maxsize", "Ungültige Eingabe für eine Ganzzahl.")
@@ -790,7 +833,7 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Werbung")
         self.traverse({'href': '/ml/mailinglist/2/change'}, )
         self.assertPresence("Admin-Team")
-        f['domain'] = const.MailinglistDomain.lists.value
+        f['domain'] = const.MailinglistDomain.lists
         self.submit(f)
         self.assertTitle("Werbung")
         self.traverse({'href': '/ml/mailinglist/2/change'}, )
@@ -855,11 +898,11 @@ class TestMlFrontend(FrontendTest):
 
         for ml_type in const.MailinglistTypes:
             with self.subTest(ml_type=ml_type):
-                f['ml_type'] = ml_type.value
+                f['ml_type'] = ml_type
                 f['event_id'] = event_id
                 f['registration_stati'] = [
-                    const.RegistrationPartStati.participant.value,
-                    const.RegistrationPartStati.waitlist.value,
+                    const.RegistrationPartStati.participant,
+                    const.RegistrationPartStati.waitlist,
                 ]
                 f['assembly_id'] = assembly_id
                 # no ml type should allow event *and* assembly fields to be set
@@ -909,12 +952,17 @@ class TestMlFrontend(FrontendTest):
         self.get("/ml/mailinglist/9/change")
         self.assertTitle("Teilnehmer-Liste – Konfiguration")
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({"2", "4", None}, tmp)
-        f['registration_stati'] = [3, 5]
+        reality = _get_registration_part_stati(f)
+        expectation = {str(const.RegistrationPartStati.participant),
+                       str(const.RegistrationPartStati.guest)}
+        self.assertEqual(expectation, reality)
+        f['registration_stati'] = [const.RegistrationPartStati.waitlist,
+                                   const.RegistrationPartStati.cancelled]
         self.submit(f)
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({"3", "5", None}, tmp)
+        reality = _get_registration_part_stati(f)
+        expectation = {str(const.RegistrationPartStati.waitlist),
+                       str(const.RegistrationPartStati.cancelled)}
+        self.assertEqual(expectation, reality)
 
     @as_users("nina")
     def test_delete_ml(self) -> None:
@@ -928,11 +976,19 @@ class TestMlFrontend(FrontendTest):
 
     def test_subscription_request(self) -> None:
         self.login(USER_DICT['inga'])
-        self.traverse({'href': '/ml/$'},
-                      {'href': '/ml/mailinglist/4'},)
+        self.traverse("Mailinglisten")
+        # check icon
+        self.assertEqual(len(self.response.lxml.get_element_by_id('mailinglist4')
+                             .find_class('fa-times-circle')), 1)
+        self.traverse("Klatsch und Tratsch")
         self.assertTitle("Klatsch und Tratsch")
         f = self.response.forms['subscribe-mod-form']
         self.submit(f)
+        self.assertIn('cancel-request-form', self.response.forms)
+        self.traverse("Mailinglisten-Übersicht")
+        # check icon
+        self.assertEqual(len(self.response.lxml.get_element_by_id('mailinglist4')
+                             .find_class('fa-circle')), 1)
         self.logout()
         self.login(USER_DICT['berta'])
         self.traverse({'href': '/ml/$'},
@@ -945,8 +1001,11 @@ class TestMlFrontend(FrontendTest):
         self.assertNotIn('handlerequestform9', self.response.forms)
         self.logout()
         self.login(USER_DICT['inga'])
-        self.traverse({'href': '/ml/$'},
-                      {'href': '/ml/mailinglist/4'},)
+        self.traverse("Mailinglisten")
+        # check icon
+        self.assertEqual(len(self.response.lxml.get_element_by_id('mailinglist4')
+                             .find_class('fa-check-circle')), 1)
+        self.traverse("Klatsch und Tratsch")
         self.assertIn('unsubscribeform', self.response.forms)
 
     @as_users("charly", "inga")
@@ -968,7 +1027,7 @@ class TestMlFrontend(FrontendTest):
         f = self.response.forms['addsubscriberform']
         f['subscriber_ids'] = "DB-1-9"
         self.submit(f)
-        self.assertPresence("Anton Armin A. Administrator")
+        self.assertPresence("Anton Administrator")
 
     def _create_mailinglist(self, mdata: CdEDBObject) -> None:
         self.traverse({'href': '/ml/'},
@@ -993,9 +1052,9 @@ class TestMlFrontend(FrontendTest):
                 self.login("anton")
                 mdata = {
                     'title': 'TestAkaList',
-                    'ml_type': const.MailinglistTypes.event_associated.value,
+                    'ml_type': const.MailinglistTypes.event_associated,
                     'local_part': 'testaka',
-                    'domain': const.MailinglistDomain.aka.value,
+                    'domain': const.MailinglistDomain.aka,
                     'event_id': "1",
                     'moderators': user['DB-ID'],
                 }
@@ -1007,7 +1066,7 @@ class TestMlFrontend(FrontendTest):
                     f = self.response.forms['addorgaform']
                     f['orga_id'] = user['DB-ID']
                     self.submit(f, check_notification=False)
-                    self.assertPresence(user['given_names'], div='manage-orgas')
+                    self.assertPresence(user['family_name'], div='manage-orgas')
                 self.logout()
                 self.login(user)
                 self.traverse({'href': '/'})
@@ -1140,13 +1199,13 @@ class TestMlFrontend(FrontendTest):
         f['event_id'].force_value(1)
         f['is_active'].force_value(False)
         # these properties can be changed by full moderators only
-        f['registration_stati'] = [const.RegistrationPartStati.guest.value]
+        f['registration_stati'] = [const.RegistrationPartStati.guest]
         # these properties can be changed by every moderator
         f['description'] = "Wir machen Party!"
         f['notes'] = "Nur geladene Gäste."
-        f['mod_policy'] = const.ModerationPolicy.unmoderated.value
+        f['mod_policy'] = const.ModerationPolicy.unmoderated
         f['subject_prefix'] = "party"
-        f['attachment_policy'] = const.AttachmentPolicy.allow.value
+        f['attachment_policy'] = const.AttachmentPolicy.allow
         f['maxsize'] = 1111
         self.submit(f)
 
@@ -1159,21 +1218,21 @@ class TestMlFrontend(FrontendTest):
         self.assertEqual(str(old_ml['event_id']), f['event_id'].value)
 
         # ... these have only changed if the moderator is privileged ...
-        reality = {f.get("registration_stati", index=i).value for i in range(7)}
+        reality = _get_registration_part_stati(f)
         if self.user_in('berta'):
-            expectation = {None, str(const.RegistrationPartStati.guest.value)}
+            expectation = {str(const.RegistrationPartStati.guest)}
         else:
-            expectation = {str(status)
-                           for status in old_ml['registration_stati']} | {None}
+            expectation = {str(const.RegistrationPartStati(status))
+                           for status in old_ml['registration_stati']}
         self.assertEqual(expectation, reality)
 
         # ... and these have changed.
         self.assertEqual("Wir machen Party!", f['description'].value)
         self.assertEqual("Nur geladene Gäste.", f['notes'].value)
-        self.assertEqual(str(const.ModerationPolicy.unmoderated.value),
+        self.assertEqual(str(const.ModerationPolicy.unmoderated),
                          f['mod_policy'].value)
         self.assertEqual("party", f['subject_prefix'].value)
-        self.assertEqual(str(const.AttachmentPolicy.allow.value),
+        self.assertEqual(str(const.AttachmentPolicy.allow),
                          f['attachment_policy'].value)
         self.assertEqual("1111", f['maxsize'].value)
 
@@ -1217,7 +1276,7 @@ class TestMlFrontend(FrontendTest):
         f = self.response.forms['addmoderatorform']
         f['moderators'] = USER_DICT['berta']['DB-ID']
         self.submit(f)
-        self.assertPresence("Bertålotta", div='moderator-list')
+        self.assertPresence("Bertå", div='moderator-list')
         self.assertPresence("Garcia", div='moderator-list')
         f = self.response.forms['removemoderatorform7']
         self.submit(f)
@@ -1273,7 +1332,7 @@ class TestMlFrontend(FrontendTest):
         self.traverse({"description": "Mailinglisten"},
                       {"description": "Mailingliste anlegen"})
         f = self.response.forms['selectmltypeform']
-        f['ml_type'] = const.MailinglistTypes.cdelokal.value
+        f['ml_type'] = const.MailinglistTypes.cdelokal
         self.assertEqual(len(f['ml_type'].options), 2)
         self.submit(f)
         f = self.response.forms['createlistform']
@@ -1283,48 +1342,49 @@ class TestMlFrontend(FrontendTest):
         f['description'] = "If anyone else lives here, please come by, " \
                            "I am lonely."
         f['local_part'] = "littlewhinging"
-        f['domain'] = const.MailinglistDomain.cdelokal.value
+        f['domain'] = const.MailinglistDomain.cdelokal
         self.assertEqual(len(f['domain'].options),
-                         len(ml_type.CdeLokalMailinglist.domains))
+                         len(CdeLokalMailinglist.domains))
         moderator = USER_DICT["berta"]
         f['moderators'] = moderator["DB-ID"]
         self.submit(f)
         self.assertTitle("Little Whinging")
-        self.assertPresence(moderator['given_names'], div="moderator-list")
+        self.assertPresence(moderator['family_name'], div="moderator-list")
 
     @as_users("anton")
     def test_1342(self) -> None:
         self.get("/ml/mailinglist/60/change")
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
+        reality = _get_registration_part_stati(f)
         sample_data_stati = set(
-            str(x) for x in self.get_sample_data(
+            str(const.RegistrationPartStati(x)) for x in self.get_sample_data(
                 "ml.mailinglists", (60,), ("registration_stati",)
             )[60]["registration_stati"])
-        self.assertEqual(sample_data_stati | {None}, tmp)
-        stati = [const.RegistrationPartStati.waitlist.value,
-                 const.RegistrationPartStati.guest.value]
+        self.assertEqual(sample_data_stati, reality)
+        stati = [const.RegistrationPartStati.waitlist,
+                 const.RegistrationPartStati.guest]
         f['registration_stati'] = stati
         self.submit(f)
         self.traverse({"description": "Konfiguration"})
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({str(x) for x in stati} | {None}, tmp)
+        reality = _get_registration_part_stati(f)
+        self.assertEqual({str(x) for x in stati}, reality)
 
-        stati = [const.RegistrationPartStati.not_applied.value]
+        stati = [const.RegistrationPartStati.not_applied]
         f['registration_stati'] = stati
         self.submit(f)
         self.traverse({"description": "Konfiguration"})
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({str(x) for x in stati} | {None}, tmp)
+        reality = _get_registration_part_stati(f)
+        self.assertEqual({str(x) for x in stati}, reality)
 
-    def _prepare_moderation_mock(self, client_class: unittest.mock.Mock) -> Tuple[
-        List[MockHeldMessage], unittest.mock.MagicMock, Any]:
+    @staticmethod
+    def _prepare_moderation_mock(client_class: unittest.mock.Mock) -> Tuple[
+            List[MockHeldMessage], unittest.mock.MagicMock, Any]:
         messages = HELD_MESSAGE_SAMPLE
         mmlist = unittest.mock.MagicMock()
         moderation_response = unittest.mock.MagicMock()
-        moderation_response.status = 204
+        moderation_response.status_code = 204
         mmlist.moderate_message.return_value = moderation_response
         client = client_class.return_value
         client.get_held_messages.return_value = messages
@@ -1380,6 +1440,76 @@ class TestMlFrontend(FrontendTest):
 
     @unittest.mock.patch("cdedb.frontend.common.CdEMailmanClient")
     @as_users("anton")
+    def test_mailman_moderation_multi_accept(self, client_class: unittest.mock.Mock
+                                             ) -> None:
+        #
+        # Prepare
+        #
+        messages, mmlist, client = self._prepare_moderation_mock(client_class)
+
+        #
+        # Run
+        #
+        self.traverse({'href': '/ml/$'},
+                      {'href': '/ml/mailinglist/99'},
+                      {'href': '/ml/mailinglist/99/moderate'})
+        self.assertTitle("Mailman-Migration – Nachrichtenmoderation")
+        self.assertPresence("Finanzbericht")
+        self.assertPresence("Verschwurbelung")
+        self.assertPresence("unerwartetes Erbe")
+        client.get_held_messages.return_value = []
+        f = self.response.forms['moderateallform']
+        self.submit(f, button='action', value='accept')
+        self.assertNonPresence("Finanzbericht")
+        self.assertNonPresence("Verschwurbelung")
+        self.assertNonPresence("unerwartetes Erbe")
+
+        #
+        # Check
+        #
+        umcall = unittest.mock.call
+        # Creation
+        self.assertEqual(
+            mmlist.moderate_message.call_args_list,
+            [umcall(1, 'accept'), umcall(2, 'accept'), umcall(3, 'accept')])
+
+    @unittest.mock.patch("cdedb.frontend.common.CdEMailmanClient")
+    @as_users("anton")
+    def test_mailman_moderation_multi_discard(self, client_class: unittest.mock.Mock
+                                              ) -> None:
+        #
+        # Prepare
+        #
+        messages, mmlist, client = self._prepare_moderation_mock(client_class)
+
+        #
+        # Run
+        #
+        self.traverse({'href': '/ml/$'},
+                      {'href': '/ml/mailinglist/99'},
+                      {'href': '/ml/mailinglist/99/moderate'})
+        self.assertTitle("Mailman-Migration – Nachrichtenmoderation")
+        self.assertPresence("Finanzbericht")
+        self.assertPresence("Verschwurbelung")
+        self.assertPresence("unerwartetes Erbe")
+        client.get_held_messages.return_value = []
+        f = self.response.forms['moderateallform']
+        self.submit(f, button='action', value='discard')
+        self.assertNonPresence("Finanzbericht")
+        self.assertNonPresence("Verschwurbelung")
+        self.assertNonPresence("unerwartetes Erbe")
+
+        #
+        # Check
+        #
+        umcall = unittest.mock.call
+        # Creation
+        self.assertEqual(
+            mmlist.moderate_message.call_args_list,
+            [umcall(1, 'discard'), umcall(2, 'discard'), umcall(3, 'discard')])
+
+    @unittest.mock.patch("cdedb.frontend.common.CdEMailmanClient")
+    @as_users("anton")
     def test_mailman_whitelist(self, client_class: unittest.mock.Mock) -> None:
         #
         # Prepare
@@ -1417,7 +1547,11 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Mailinglisten-Log [1–9 von 9]")
         self.assertNonPresence("LogCodes")
         f = self.response.forms['logshowform']
-        f['codes'] = [10, 11, 20, 21, 22]
+        f['codes'] = [const.MlLogCodes.moderator_added,
+                      const.MlLogCodes.moderator_removed,
+                      const.MlLogCodes.subscription_requested,
+                      const.MlLogCodes.subscribed,
+                      const.MlLogCodes.subscription_changed]
         f['mailinglist_id'] = 4
 
         self.submit(f)
